# docs/swagger.yaml
definitions:
  graphdiff.Change:
    properties:
      change:
        description: '"added" | "removed" | "modified"'
        type: string
      fields:
        items:
          $ref: '#/definitions/graphdiff.FieldChange'
        type: array
      orbId:
        type: string
      type:
        type: string
    type: object
  graphdiff.FieldChange:
    properties:
      after: {}
      before: {}
      field:
        type: string
    type: object
  graphdiff.Summary:
    properties:
      added:
        type: integer
      modified:
        type: integer
      removed:
        type: integer
      unchanged:
        type: integer
    type: object
  handler.artifactRef:
    properties:
      dataCenterName:
        example: colo-galleon
        type: string
      digest:
        example: sha256:81147a187a8f
        type: string
      id:
        example: 42
        type: integer
      publishedAt:
        type: string
      tag:
        example: v2
        type: string
    type: object
...
  handler.compareResponse:
    properties:
      byType:
        additionalProperties:
          $ref: '#/definitions/graphdiff.Summary'
        type: object
      changes:
        items:
          $ref: '#/definitions/graphdiff.Change'
        type: array
      disclaimer:
        type: string
      from:
        $ref: '#/definitions/handler.artifactRef'
      summary:
        $ref: '#/definitions/graphdiff.Summary'
      to:
        $ref: '#/definitions/handler.artifactRef'
    type: object
  handler.currentMeta:
    properties:
      contentHash:
        example: sha256:abc
        type: string
      nodeCount:
        example: 663
        type: integer
      source:
        example: blue
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      code:
        example: FORBIDDEN
        type: string
      docUrl:
        type: string
      error:
        example: dev or admin role required for mutations
        type: string
      hint:
        example: Ask an admin to grant you the dev role.
        type: string
      httpStatus:
        example: 403
        type: integer
    type: object
...
  handler.exportRequest:
    properties:
      download:
        description: |-
          Download selects the mode. false (default) exports and publishes to OCI,
          discarding the zip on completion. true exports only and retains the zip
          for GET /api/v1/export/jobs/{jobId}/download.
        example: false
        type: boolean
      expectedContentHash:
        description: |-
          ExpectedContentHash is the optional guarded-Apply token: pass the
          `current.contentHash` from a prior POST /api/v1/export/preview and the
          export aborts with 409 if the data center's desired state changed since
          that preview (another writer edited intent in the review→Apply gap).
          Omit for last-writer-wins (default, back-compatible). This is the
          ETag/If-Match pattern applied to a whole subgraph — one hash rather than
          per-node version checks, so concurrent ADDS and DELETES are caught too.
        example: sha256:045b8a51a0aea59fa
        type: string
      orbId:
        description: OrbID identifies the data center to export, in "namespace:name"
          form. Required.
        example: alaska:dc-01
        type: string
    required:
    - orbId
    type: object
...
  handler.lastPublishedMeta:
    properties:
      digest:
        example: sha256:1a2b
        type: string
      exportJobId:
        type: string
      publishedAt:
        type: string
      reason:
        type: string
      retrievable:
        type: boolean
      state:
        description: published | first_export | unavailable
        example: published
        type: string
      tag:
        example: v8
        type: string
    type: object
  handler.previewResponse:
    properties:
      byType:
        additionalProperties:
          $ref: '#/definitions/graphdiff.Summary'
        type: object
      changes:
        items:
          $ref: '#/definitions/graphdiff.Change'
        type: array
      current:
        $ref: '#/definitions/handler.currentMeta'
      dataCenterName:
        example: colo-galleon
        type: string
      disclaimer:
        type: string
      lastPublishedVersion:
        $ref: '#/definitions/handler.lastPublishedMeta'
      orbId:
        example: colo:colo-galleon
        type: string
      summary:
        $ref: '#/definitions/graphdiff.Summary'
    type: object
...
  handler.triggerResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API-first, graph-native source of truth for modular data centers.
  title: Orbital API
  version: 0.0.0-dev
paths:
...
  /api/v1/export:
    post:
      consumes:
      - application/json
      description: Triggers an async atomic export of the data center's configuration
        subgraph. When download=false (default) AND OCI is configured, the export
        chains into an OCI publish and the zip is discarded on completion. When download=true
        (or OCI is not configured), the export runs in download-only mode and the
        zip is retained on disk for the client to fetch via GET /api/v1/export/jobs/{jobId}/download.
        Returns 202 with a job ID immediately; poll GET /api/v1/export/jobs/{jobId}
        until terminal state. Returns 409 if an export or restore is already in progress,
        or if expectedContentHash is supplied and the desired state changed since
        that preview (guarded Apply).
      parameters:
      - description: Export request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.exportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.triggerResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Trigger subgraph export
      tags:
      - export
  /api/v1/export/compare:
    get:
      description: Returns the desired-state delta between two published OCI artifacts
        of the same data center. Read-only; both artifacts are pulled by immutable
        digest. Not an apply-forecast.
      parameters:
      - description: Artifact ID to diff FROM (the earlier side)
        in: query
        name: from
        required: true
        type: integer
      - description: Artifact ID to diff TO (the later side)
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.compareResponse'
      summary: Compare two published artifacts
      tags:
      - export
...
  /api/v1/export/preview:
    post:
      consumes:
      - application/json
      description: Synchronous, read-only. Diffs the current desired state for a data
        center against its last published artifact. Not an apply-forecast.
      parameters:
      - description: Data center orbId to preview
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.exportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.previewResponse'
      summary: Preview export diff (desired-state delta vs last published)
      tags:
      - export
...
swagger: "2.0"
...